ExperimentsExample/train/requirements.txt
pandas
sensa-python==6.5.0
scikit-learn==0.24.2
numpy
boto3
